move regionmapping imagery provider to another file
