+++ lib/ModelMixins/TableMixin.ts
@@ -46,6 +46,7 @@ import DiscretelyTimeVaryingMixin, {
 } from "./DiscretelyTimeVaryingMixin";
 import ExportableMixin, { ExportData } from "./ExportableMixin";
 import { ImageryParts } from "./MappableMixin";
+import createRegionMappedImageryProvider from "../Table/createRegionMappedImageryProvider";
 
 function TableMixin<T extends Constructor<Model<TableTraits>>>(Base: T) {
   abstract class TableMixin
@@ -702,225 +703,9 @@ function TableMixin<T extends Constructor<Model<TableTraits>>>(Base: T) {
       (input: {
         style: TableStyle;
         currentTime: JulianDate | undefined;
-      }): ImageryProvider | undefined => {
-        if (!input.style.isRegions()) {
-          return undefined;
-        }
-
-        const regionColumn = input.style.regionColumn;
-        const regionType = regionColumn.regionType;
-        if (regionType === undefined) {
-          return undefined;
-        }
-
-        const baseMapContrastColor = "white"; //this.terria.baseMapContrastColor;
-
-        const colorColumn = input.style.colorColumn;
-        const valueFunction =
-          colorColumn !== undefined
-            ? colorColumn.valueFunctionForType
-            : () => null;
-        const colorMap = (this.activeTableStyle || this.defaultTableStyle)
-          .colorMap;
-        const valuesAsRegions = regionColumn.valuesAsRegions;
-
-        let currentTimeRows: number[] | undefined;
-
-        // If time varying, get row indices which match
-        if (
-          input.currentTime &&
-          input.style.timeIntervals &&
-          input.style.moreThanOneTimeInterval
-        ) {
-          currentTimeRows = input.style.timeIntervals.reduce<number[]>(
-            (rows, timeInterval, index) => {
-              if (
-                timeInterval &&
-                TimeInterval.contains(timeInterval, input.currentTime!)
-              ) {
-                rows.push(index);
-              }
-              return rows;
-            },
-            []
-          );
-        }
-
-        const catalogItem = this;
-
-        return new MapboxVectorTileImageryProvider({
-          url: regionType.server,
-          layerName: regionType.layerName,
-          styleFunc: function(feature: any) {
-            const regionId = feature.properties[regionType.uniqueIdProp];
-
-            let rowNumber = catalogItem.getImageryLayerFilteredRows(
-              input,
-              currentTimeRows,
-              valuesAsRegions.regionIdToRowNumbersMap.get(regionId)
-            );
-            let value: string | number | null = isDefined(rowNumber)
-              ? valueFunction(rowNumber)
-              : null;
-
-            const color = colorMap.mapValueToColor(value);
-            if (color === undefined) {
-              return undefined;
-            }
-
-            return {
-              fillStyle: color.toCssColorString(),
-              strokeStyle: baseMapContrastColor,
-              lineWidth: 1,
-              lineJoin: "miter"
-            };
-          },
-          subdomains: regionType.serverSubdomains,
-          rectangle:
-            Array.isArray(regionType.bbox) && regionType.bbox.length >= 4
-              ? Rectangle.fromDegrees(
-                  regionType.bbox[0],
-                  regionType.bbox[1],
-                  regionType.bbox[2],
-                  regionType.bbox[3]
-                )
-              : undefined,
-          minimumZoom: regionType.serverMinZoom,
-          maximumNativeZoom: regionType.serverMaxNativeZoom,
-          maximumZoom: regionType.serverMaxZoom,
-          uniqueIdProp: regionType.uniqueIdProp,
-          featureInfoFunc: (feature: any) =>
-            this.getImageryLayerFeatureInfo(input, feature, currentTimeRows)
-        });
-      }
+      }): ImageryProvider | undefined =>
+        createRegionMappedImageryProvider(input.style, input.currentTime)
     );
-
-    /**
-     * Filters row numbers by time (if applicable) - for a given region mapped ImageryLayer
-     */
-    private getImageryLayerFilteredRows(
-      input: {
-        style: TableStyle;
-        currentTime: JulianDate | undefined;
-      },
-      currentTimeRows: number[] | undefined,
-      rowNumbers: number | readonly number[] | undefined
-    ): number | undefined {
-      if (!isDefined(rowNumbers)) return;
-
-      if (!isDefined(currentTimeRows)) {
-        return Array.isArray(rowNumbers) ? rowNumbers[0] : rowNumbers;
-      }
-
-      if (
-        typeof rowNumbers === "number" &&
-        currentTimeRows.includes(rowNumbers)
-      ) {
-        return rowNumbers;
-      } else if (Array.isArray(rowNumbers)) {
-        const matchingTimeRows: number[] = rowNumbers.filter(row =>
-          currentTimeRows!.includes(row)
-        );
-        if (matchingTimeRows.length <= 1) {
-          return matchingTimeRows[0];
-        }
-        // In a time-varying dataset, intervals may
-        // overlap at their endpoints (i.e. the end of one interval is the start of the next).
-        // In that case, we want the later interval to apply.
-        return matchingTimeRows.reduce((latestRow, currentRow) => {
-          const currentInterval = input.style.timeIntervals?.[currentRow]?.stop;
-          const latestInterval = input.style.timeIntervals?.[latestRow]?.stop;
-          if (
-            currentInterval &&
-            latestInterval &&
-            JulianDate.lessThan(latestInterval, currentInterval)
-          ) {
-            return currentRow;
-          }
-          return latestRow;
-        }, matchingTimeRows[0]);
-      }
-    }
-
-    /**
-     * Get ImageryLayerFeatureInfo for a given ImageryLayer input and feature.
-     */
-    private getImageryLayerFeatureInfo(
-      input: {
-        style: TableStyle;
-        currentTime: JulianDate | undefined;
-      },
-      feature: VectorTileFeature,
-      currentTimeRows: number[] | undefined
-    ) {
-      if (
-        isDefined(input.style.regionColumn) &&
-        isDefined(input.style.regionColumn.regionType) &&
-        isDefined(input.style.regionColumn.regionType.uniqueIdProp)
-      ) {
-        const regionType = input.style.regionColumn.regionType;
-
-        if (!isDefined(regionType)) return undefined;
-
-        const regionIds =
-          input.style.regionColumn.valuesAsRegions.regionIdToRowNumbersMap.get(
-            feature.properties[regionType.uniqueIdProp]
-          ) ?? [];
-
-        const filteredRowIds = this.getImageryLayerFilteredRows(
-          input,
-          currentTimeRows,
-          regionIds
-        );
-
-        let d: JsonObject | null = isDefined(filteredRowIds)
-          ? this.getRowValues(filteredRowIds)
-          : null;
-
-        if (d === null) return;
-
-        // Preserve values from d and insert feature properties after entries from d
-        const featureData = Object.assign({}, d, feature.properties, d);
-
-        const featureInfo = new ImageryLayerFeatureInfo();
-        if (isDefined(regionType.nameProp)) {
-          featureInfo.name = featureData[regionType.nameProp] as string;
-        }
-
-        featureData.id = feature.properties[regionType.uniqueIdProp];
-        featureInfo.properties = featureData;
-
-        featureInfo.configureDescriptionFromProperties(featureData);
-        featureInfo.configureNameFromProperties(featureData);
-
-        // If time-series region-mapping - show timeseries chart
-        if (
-          !isDefined(featureData._terria_getChartDetails) &&
-          this.discreteTimes &&
-          this.discreteTimes.length > 1 &&
-          Array.isArray(regionIds)
-        ) {
-          featureInfo.properties._terria_getChartDetails = getChartDetailsFn(
-            this.activeTableStyle,
-            regionIds
-          );
-        }
-
-        return featureInfo;
-      }
-
-      return undefined;
-    }
-
-    private getRowValues(index: number): JsonObject {
-      const result: JsonObject = {};
-
-      this.tableColumns.forEach(column => {
-        result[column.title] = column.valueFunctionForType(index);
-      });
-
-      return result;
-    }
 
     private readonly getTableColumn: ITransformer<
       number,
